If else заменен тернарным оператором, класс и интерфейс вынесены в отдельный файл
