+++ src/main/java/faang/school/godbless/lord_of_the_rings/Character.java
@@ -1,0 +1,13 @@
+package faang.school.godbless.lord_of_the_rings;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+import java.util.List;
+
+@AllArgsConstructor
+@Getter
+class Character {
+    private String name;
+    private List<Item> inventory;
+}
