+++ src/main/java/faang/school/godbless/lord_of_the_rings/InventoryManager.java
@@ -1,9 +1,5 @@
 package faang.school.godbless.lord_of_the_rings;
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-import java.util.List;
 import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.function.Predicate;
@@ -20,24 +16,7 @@ public class InventoryManager {
 
     public static void updateItem(Character character, Predicate<Item> predicate, Function<Item, Item> function) {
         character.getInventory().replaceAll(item -> {
-            if (predicate.test(item)) {
-                return function.apply(item);
-            }
-            return item;
+            return predicate.test(item) ? function.apply(item) : item;
         });
     }
-}
-
-@AllArgsConstructor
-@Getter
-class Character {
-    private String name;
-    private List<Item> inventory;
-}
-
-@AllArgsConstructor
-@Getter
-class Item {
-    private String name;
-    private int value;
 }
