+++ src/main/java/faang/school/godbless/lord_of_the_rings/Item.java
@@ -1,0 +1,11 @@
+package faang.school.godbless.lord_of_the_rings;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+class Item {
+    private String name;
+    private int value;
+}
